Version33

+++ app/src/androidTest/kotlin/com/mstoyanov/musiclessons/MigrationTest.kt
@@ -10,7 +10,6 @@ import android.database.sqlite.SQLiteOpenHelper
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import com.mstoyanov.musiclessons.repository.AppDatabase
-import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -23,30 +22,27 @@ class MigrationTest {
     // this is a Room persistence helper:
     val helper: MigrationTestHelper = MigrationTestHelper(InstrumentationRegistry.getInstrumentation(), AppDatabase::class.java.canonicalName, FrameworkSQLiteOpenHelperFactory())
 
-    @Before
+    @Test
     @Throws(IOException::class)
-    fun setup() {
+    fun migrate_1_to_3_test() {
         // this is an SQLite helper:
-        val helper = TestSQLiteOpenHelper(InstrumentationRegistry.getTargetContext())
-        val db = helper.writableDatabase
+        val sqLiteHelper = TestSQLiteOpenHelper(InstrumentationRegistry.getTargetContext())
+        val db = sqLiteHelper.writableDatabase
 
         db.execSQL("create table if not exists students (studentID integer primary key autoincrement not null, firstName text, lastName text, homePhone text, cellPhone text, workPhone text, email text);")
         db.execSQL("create table if not exists schedule (lessonID integer primary key autoincrement not null, weekday text, timeFrom text, timeTo text, studentID integer);")
 
         db.execSQL("insert into students (firstName, lastName, homePhone, cellPhone, workPhone, email) values ('Aaaa', 'Bbbb', '123-456-7890', '456-789-0000', '789-000-1234', 'aaabbb@gmail.com');")
-        db.execSQL("insert into students (firstName, homePhone, cellPhone) values ('Cccc', '000-123-4444', '111-222-3456');")
-        db.execSQL("insert into students (firstName, homePhone) values ('Dddd', '');")
+        db.execSQL("insert into students (firstName, lastName, homePhone, cellPhone) values ('Cccc', 'Cc', '000-123-4444', '111-222-3456');")
+        db.execSQL("insert into students (firstName, lastName, homePhone) values ('Dddd', 'Dd', '222-3456-7777');")
+
         db.execSQL("insert into schedule (weekday, timeFrom, timeTo, studentID) values ('Monday', '15:30', '16:00', '1');")
         db.execSQL("insert into schedule (weekday, timeFrom, timeTo, studentID) values ('Monday', '16:00', '16:30', '2');")
         db.execSQL("insert into schedule (weekday, timeFrom, timeTo, studentID) values ('Monday', '16:30', '17:00', '3');")
         db.execSQL("insert into schedule (weekday, timeFrom, timeTo, studentID) values ('Tuesday', '15:30', '16:00', '1');")
 
         db.close()
-    }
 
-    @Test
-    @Throws(IOException::class)
-    fun migrate_1_to_3_test() {
         helper.runMigrationsAndValidate(TEST_DB, 3, true, MIGRATION_1_3)
     }
 
@@ -57,12 +53,14 @@ class MigrationTest {
 
         db.execSQL("insert into student (first_name, last_name, email, notes) values ('Aaaa', 'Bbbb', 'aaabbb@gmail.com', 'notes');")
         db.execSQL("insert into student (first_name) values ('Cccc');")
-        db.execSQL("insert into student (first_name) values ('Dddd');")
+        db.execSQL("insert into student (last_name) values ('Dddd');")
+
         db.execSQL("insert into phone_number (number, student_id, type) values ('123-456-7890', 1, 'Home');")
         db.execSQL("insert into phone_number (number, student_id, type) values ('456-789-0000', 1, 'Cell');")
         db.execSQL("insert into phone_number (number, student_id, type) values ('789-000-1234', 1, 'Work');")
         db.execSQL("insert into phone_number (number, student_id, type) values ('000-123-4444', 2, 'Home');")
         db.execSQL("insert into phone_number (number, student_id, type) values ('111-222-3456', 2, 'Cell');")
+
         db.execSQL("insert into lesson (weekday, student_id, time_from, time_to) values('Monday', 1, (select strftime('%s', '1970-01-01 15:30') * 1000), (select strftime('%s', '1970-01-01 16:00') * 1000));")
         db.execSQL("insert into lesson (weekday, student_id, time_from, time_to) values('Monday', 2, (select strftime('%s', '1970-01-01 16:00') * 1000), (select strftime('%s', '1970-01-01 16:30') * 1000));")
         db.execSQL("insert into lesson (weekday, student_id, time_from, time_to) values('Monday', 3, (select strftime('%s', '1970-01-01 16:30') * 1000), (select strftime('%s', '1970-01-01 17:00') * 1000));")
@@ -84,7 +82,7 @@ class MigrationTest {
                 database.execSQL("create table if not exists lesson (lesson_id integer primary key autoincrement not null, weekday text not null, time_from integer not null, time_to integer not null, student_id integer not null, foreign key(student_id) references student(s_id) on update no action on delete cascade);")
                 database.execSQL("create index index_lesson_student_id on lesson (student_id);")
 
-                database.execSQL("insert into student (first_name, last_name, email, notes) select case when firstName is null then '' else firstName end, case when lastName is null then '' else lastName end, case when email is null then '' else email end, '' from students;")
+                database.execSQL("insert into student (first_name, last_name, email, notes) select firstName, lastName, case when email is null then '' else email end, '' from students;")
                 database.execSQL("insert into phone_number (number, student_id, type) select homePhone, studentID, 'Home' from students where homePhone is not null and length(homePhone) > 0;")
                 database.execSQL("insert into phone_number (number, student_id, type) select cellPhone, studentID, 'Cell' from students where cellPhone is not null and length(cellPhone) > 0;")
                 database.execSQL("insert into phone_number (number, student_id, type) select workPhone, studentID, 'Work' from students where workPhone is not null and length(workPhone) > 0;")
@@ -104,7 +102,7 @@ class MigrationTest {
                 database.execSQL("create index index_lesson3_student_id on lesson3 (student_id);")
 
                 database.execSQL("insert into student3 (first_name, last_name, email, notes) select case when first_name is null then '' else first_name end, case when last_name is null then '' else last_name end, case when email is null then '' else email end, case when notes is null then '' else notes end from student;")
-                database.execSQL("insert into phone_number3 (number, student_id,type) select case when number is null then '' else number end, case when student_id is null then '' else student_id end, case when type is null then '' else type end from phone_number;")
+                database.execSQL("insert into phone_number3 (number, student_id, type) select number, student_id, type from phone_number;")
                 database.execSQL("insert into lesson3 (weekday, student_id, time_from, time_to) select weekday, student_id,  time_from, time_to from lesson;")
 
                 database.execSQL("drop table if exists student;")
@@ -129,9 +127,11 @@ class MigrationTest {
         }
 
         override fun onDowngrade(sqLiteDatabase: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-            sqLiteDatabase.execSQL("drop table if exists student;")
-            sqLiteDatabase.execSQL("drop table if exists phone_number;")
-            sqLiteDatabase.execSQL("drop table if exists lesson;")
+            if (oldVersion == 1) {
+                sqLiteDatabase.execSQL("drop table if exists student;")
+                sqLiteDatabase.execSQL("drop table if exists phone_number;")
+                sqLiteDatabase.execSQL("drop table if exists lesson;")
+            }
         }
     }
 }
